refactor: use same template structure as for Actor templates
